requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
motor>=3.3.1
pymongo>=4.6.0
dnspython>=2.4.2